% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Adds data field to metadata file.}
\usage{
add_metadata(..., path = here::here("metadata.csv"))
}
\arguments{
\item{...}{Additional parameters. Supports: Column names.}

\item{path}{str | CSV file containing metadata (default: "metadata.csv").}
}
\value{
NA
}
\description{
Adds data field to metadata file.
}
\examples{
# Create metadata
create_metadata("constant",
                "column_name",
                "display_name",
                "col_type")

# Add metadata
add_metadata(constant = BMI,
             column_name = BMI,
             display_name = "Body mass index (BMI)",
             col_type = "dbl")
}
