% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjSigVarExp.R
\name{AdjSigVarExp}
\alias{AdjSigVarExp}
\title{Adjust Dataset Components to get Desired R^2 Values}
\usage{
AdjSigVarExp(J, I, E, JntVarEx, IndVarEx)
}
\arguments{
\item{J}{The joint component of the dataset (a matrix).}

\item{I}{The individual component of the dataset (a matrix).}

\item{E}{The noise component of the dataset (a matrix).}

\item{JntVarEx}{Proportions of joint variation explained (input as a vector).}

\item{IndVarEx}{Proportions of individual variation explained (input as a vector).}
}
\value{
A list containing:
- J: The adjusted joint component matrix.
- I: The adjusted individual component matrix.
- Data: The new dataset matrix after adjustment.
}
\description{
Adjust Dataset Components to get Desired R^2 Values
}
\examples{
J <- matrix(rnorm(100), 10, 10)
I <- matrix(rnorm(100), 10, 10)
E <- matrix(rnorm(100), 10, 10)
result <- AdjSigVarExp(J, I, E, JntVarEx = 0.3, IndVarEx = 0.5)
str(result)
}
