% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{archive_con}
\alias{archive_con}
\title{Construct a new connection into an archive}
\usage{
archive_con(archive, file = 1L)
}
\arguments{
\item{archive}{An archive object or character vector to the archive}

\item{file}{The file to open the connection to. Can also be an numeric index
into the `archive()` data.frame.}
}
\description{
Construct a new connection into an archive
}
