% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{archive_extract}
\alias{archive_extract}
\title{Extract contents of an archive}
\usage{
archive_extract(archive, dir = ".", file = NULL)
}
\arguments{
\item{archive}{`character(1)` The archive filename or an `archive` object.}

\item{dir}{`character(1)` Directory location to extract archive contents, will be created
if it does not exist.}

\item{file}{`character(1) || integer(1)` The filename within the archive, specified either by filename or by position.}
}
\description{
Extract contents of an archive
}
\details{
If `file` is `NULL` (the default) all files will be extracted.
}
\examples{
a <- archive(system.file(package = "archive", "extdata", "data.zip"))
d <- tempfile()
dir.create(d)
archive_extract(a, d)
list.files(d)
unlink(d)
}
