% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{archive_read}
\alias{archive_read}
\title{Construct a read only connection into an archive file}
\usage{
archive_read(archive, file = 1L, mode = "r", format = NULL,
  filter = NULL)
}
\arguments{
\item{archive}{An archive object or character vector to the archive}

\item{file}{The file to open the connection to. Can also be an numeric index
into the `archive()` data.frame.}

\item{mode}{The mode to open the file in.}

\item{format}{`character(1)` default: `NULL` The archive format, one of \Sexpr[stage=render, results=rd]{archive:::choices_rd(names(archive:::archive_formats()))}.}

\item{filter}{`character(1)` default: `NULL` The archive filter, one of \Sexpr[stage=render, results=rd]{archive:::choices_rd(names(archive:::archive_filters()))}.}
}
\description{
Construct a read only connection into an archive file
}
\details{
libarchive versions prior to 3.1.0 did not support explicit setting of the
format and filter, instead relying on a bidding process to automatically
determine the format of the archive.
}
