% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{archive_write}
\alias{archive_write}
\title{Construct a write only connection into an archive}
\usage{
archive_write(archive, file)
}
\arguments{
\item{archive}{The archive filename}

\item{file}{The filename within the archive.}
}
\description{
Construct a write only connection into an archive
}
