% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pheno_and_scan1rqtl2.R
\name{calc_pheno_vals}
\alias{calc_pheno_vals}
\title{Cacluate STR mutator phenotypes}
\usage{
calc_pheno_vals(
  in_df,
  strains_per_loc,
  pheno,
  min_pts_per_phe,
  max_denovo_strains_per_loc,
  ml,
  gtloc_per_strain = NULL
)
}
\arguments{
\item{in_df}{TODO}

\item{strains_per_loc}{TODO}

\item{pheno}{TODO}

\item{min_pts_per_phe}{TODO}

\item{max_denovo_strains_per_loc}{TODO}

\item{ml}{TODO: motif length}

\item{gtloc_per_strain}{TODO}
}
\value{
tibble
}
\description{
Function for calculating various STR mutator phenotypes which could be used for QTL mapping
}
