% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helpers_v3.R
\name{.fit_Q_ml}
\alias{.fit_Q_ml}
\title{Fit Q model (classification) on training data
Uses ranger if available, otherwise enriched GLM}
\usage{
.fit_Q_ml(train_dat)
}
\description{
Fit Q model (classification) on training data
Uses ranger if available, otherwise enriched GLM
}
