% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extracts metadata}
\usage{
extract_metadata(columns, path = here::here("metadata.csv"), col_type = NULL)
}
\arguments{
\item{columns}{str | Column or vector of columns you want to extract}

\item{path}{str | Parh to metadata. Default: "here::here("metadata.csv)"}

\item{col_type}{str | Type of data. Default: "NULL"}
}
\value{
Tibble with extracted metadata.
}
\description{
Extracts metadata
}
\examples{
# Create metadata
create_metadata("constant",
                "column_name",
                "display_name",
                "col_type")

# Add metadata
add_metadata(constant = BMI,
             column_name = BMI,
             display_name = "Body mass index (BMI)",
             col_type = "dbl")

# Extract metadata
extract_metadata("BMI")
}
