% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{file_write}
\alias{file_write}
\title{Construct a write only connection to a (possibly compressed) file.}
\usage{
file_write(file, type)
}
\arguments{
\item{file}{The filename within the archive.}

\item{type}{The type of filters to use for the file.}
}
\description{
This works similar to R's builtin [connections]. However it supports one on
or more of.
}
\details{
- uuencode
- gzip compression
- bzip2 compression
- compress/LZW compression
- lzma, lzip, and xz compression
- lz4 compression
- lzop compression
}
