% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{mode.value}
\alias{mode.value}
\title{Determine the mode of any given vector \code{x}}
\usage{
mode.value(x)
}
\arguments{
\item{x}{vector of values; can be integers/strings/booleans/etc.}
}
\value{
value the value that occurs most often
}
\description{
It's kinda dumb that R doesn't have a built-in mode function. So here.
}
\examples{
# Construct vector where 10 appears three times.
x <- c(1:10, 10:20, 10:30)
mode.value(x)

}
