% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{\code{plot_box} generates a box with an "X" in its top-right corner 
and places 1 or more text strings (of a character vector \code{lbls}) 
in the box.}
\usage{
plot_box(lbls = NA, col_lbl = "white",
  col_bg = as.character(seeblau), cex = 2, font = 2)
}
\arguments{
\item{lbls}{A character vector specifying the text label(s)
to be written. If \code{lbls} contains multiple character strings, 
each one is placed on a new line. 
Default: \code{lbls = NA}.}

\item{col_lbl}{The color(s) of the text label(s).
Default: \code{col_lbl = "white"}.}

\item{col_bg}{The color(s) to fill the box with. 
Default: \code{col_bg = seeblau}.}

\item{cex}{A numeric character expansion factor, which is
multiplied by \code{par("cex")} to yield the final character size. 
Default: \code{cex = 2}.}

\item{font}{The font to be used. 
Default: \code{font = 2} (bold).}
}
\description{
\code{plot_box} uses the base graphics system \code{graphics::}.
}
\examples{
plot_box(lbls = "A heading appears here.")
plot_box(lbls = c("Some title", "The second line is longer", "A third short line"), cex = 2.4)

plot_box(lbls = "The darkest shade of seeblau.", col_bg = pal_seeblau[[5]], cex = 1.5)

# Note: As a vector of character strings for lbls is converted into separate lines of text,
#       the following examples yield identical results:
plot_box(lbls = c("The 1st line of text.", "A 2nd and longer line of text.", "The 3rd line of text."))
plot_box(lbls = "The 1st line of text.\\nA 2nd and longer line of text.\\nThe 3rd line of text.")

# Box logos:
plot_box(lbls = "R", col_bg = pal_seeblau[[5]], cex = 13)
plot_box(lbls = c("unikn::"), col_bg = pal_seeblau[[3]], cex = 3.5)

        
}
