% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binspp_auxiliary.R
\name{print.outputs}
\alias{print.outputs}
\title{Text output describing the posterior distributions}
\usage{
\method{print}{outputs}(Output)
}
\arguments{
\item{Output}{list, output of the main function binspp.}
}
\value{
Text output summarizing the posterior distributions.
}
\description{
This function provides the summaries of the posterior distributions in the text form.
}
\examples{

# Prepare the dataset:
X = trees_N4
x_left = x_left_N4
x_right = x_right_N4
y_bottom = y_bottom_N4
y_top = y_top_N4

z_beta = list(refor=cov.refor, slope=cov.slope)
z_alpha = list(tmi=cov.tmi, tdensity=cov.tdensity)
z_omega = list(slope=cov.slope, reserv=cov.reserv)

# Determine the union of rectangles:
W = owin(c(x_left[1],x_right[1]),c(y_bottom[1],y_top[1]))
if(length(x_left)>=2){
  for(i in 2:length(x_left)){
    W2 = owin(c(x_left[i],x_right[i]),c(y_bottom[i],y_top[i]))
    W=union.owin(W,W2)
  }
}

# Dilated observation window:
W_dil = dilation.owin(W,100)


# Default parameters for prior distributions:
control = list(NStep=500, BurnIn=100, SamplingFreq=10)


# MCMC estimation:
Output = binspp(X, control, x_left, x_right, y_bottom, y_top, W_dil, z_beta, z_alpha, z_omega)


# Text output + series of figures:
print.outputs(Output)
plot.outputs(Output)

}
