% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProJIVE_BootsratVar.R
\name{ProJIVE_BootsratVar}
\alias{ProJIVE_BootsratVar}
\title{Bootstrap}
\usage{
ProJIVE_BootsratVar(B = 50, P, Q, theta.hat, W.hat, error.vars)
}
\arguments{
\item{B}{Integer. The number of bootstrap samples desired to generate (default = 50).}

\item{P}{A integer vector that contains the number of variables in each dataset.}

\item{Q}{A integer vector that contains the number of components in each dataset.}

\item{theta.hat}{Matrix. The estimated subject scores from the ProJIVE model.}

\item{W.hat}{Matrix. The estimated loading matrix from the ProJIVE model.}

\item{error.vars}{A numeric vector containing the estimated noise variances for each dataset.}
}
\value{
A list containing:
- Bootstrap_Covariance: The covariance matrix of the bootstrapped loadings.
- Bootstrap_StdErrs: The standard errors of the loadings computed from bootstrap variance.
}
\description{
Bootstrap
}
\examples{
set.seed(123)
P <- c(10, 15)
Q <- c(3, 4)
theta.hat <- matrix(rnorm(200), 20, 10)
W.hat <- matrix(rnorm(100), 10, 10)
error.vars <- c(0.5, 0.7)
result <- ProJIVE_BootsratVar(B, P, Q, theta.hat, W.hat, error.vars)
str(result)
}
