% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark-arts.R
\name{read.bq.table}
\alias{read.bq.table}
\title{Convert a copy + pasted BigQuery output table into a tibble for sharing.}
\usage{
read.bq.table(table.text)
}
\arguments{
\item{table.text}{the copied text output from a BQ query}
}
\description{
I'm not proud of this function, but I'd be lying if I said I didn't use it.
To use it, first run any query in BigQuery. Once the results are returned,
you can just copy and paste the table output into a script or console, surround it with
quotes, and pipe it into this function.
}
