% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-processing.R
\name{remove.zero.var.factors}
\alias{remove.zero.var.factors}
\title{Identify, alert, and remove columns with near zero variance.}
\usage{
remove.zero.var.factors(data)
}
\arguments{
\item{data}{any dataframe-esque object containing a mix of factor and numerical columns.}
}
\description{
This function is yet another fairly thin wrapper around \code{\link[caret]{nearZeroVar}}.
caret's function returns a vector of indices that are near zero variance, and this function
goes further to (a) inform the analyst of what columns are being removed and (b) removing
those columns.
}
\examples{

x <- data.frame(
  id     = seq(1, 10),
  age    = rpois(n=10, lambda=20),
  gender = rep(c('M', 'F'), 5),
  blood.type = factor(rep(c('A', 'B'), 5)),
  debt = factor(1e6)
)

# Removes only `debt` because (1) it's a factor and (2) has zero variance.
remove.zero.var.factors(x)

}
