% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{repr_probes}
\alias{repr_probes}
\title{Representative probes}
\format{
a data frame with:
\describe{
\item{dset}{Dataset id; e.g. GN105}
\item{probe}{Probe name; e.g. A_51_P249793}
\item{probe_chr}{Chromosome name for probe}
\item{gene_symbol}{Gene symbol that probe tags; e.g. 0710001B24Rik}
}
}
\source{

}
\usage{
repr_probes
}
\description{
List of representative probes per GeneNetwork gene per dataset in case there are
multiple probes per gene
}
\keyword{datasets}
