% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{sym.diff}
\alias{sym.diff}
\title{Compute the symmetric difference between two vectors.}
\usage{
sym.diff(x, y, pretty = FALSE)
}
\arguments{
\item{x}{one of the two vectors (soon-to-be sets) of interest}

\item{y}{one of the two vectors (soon-to-be sets) of interest}

\item{pretty}{boolean denoting whether we should also output the one- and two-sided set differences.
More appropriate for interactive analyses.}
}
\description{
R's base setdiff function does a one-way set difference, e.g. setdiff(x, y) = x - y.
More often than not, I'm wondering what values are different between x and y -- both ways.
Hence, symmetric difference. Full disclosure: I copied the core of this function from the
following StackOverflow post, with some basic tidying for interactive analysis.
}
\details{
Source: https://stackoverflow.com/questions/19797954/function-to-find-symmetric-difference-opposite-of-intersection-in-r
}
\examples{
# Find out what differences there are between 1:10 and 5:15
sym.diff(1:10, 5:15) # `1, 2, 3, 4, 11, 12, 13, 14, 15`

# Verbose output version:
sym.diff(1:10, 5:15, pretty = TRUE) # Output one- and two-sided set difference
}
