% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\name{write_internal_data}
\alias{write_internal_data}
\title{Writes metadata file into internal data.}
\usage{
write_internal_data(path = here::here("metadata.csv"))
}
\arguments{
\item{path}{str | Path to metadata file. Default: "metadata.csv".}
}
\value{
NA
}
\description{
Writes metadata file into internal data.
}
\examples{
# Create metadata
create_metadata("constant",
                "column_name",
                "display_name",
                "col_type")

# Add metadata
add_metadata(constant = BMI,
             column_name = BMI,
             display_name = "Body mass index (BMI)",
             col_type = "dbl")

write_internal_data()
}
