% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{archive}
\alias{archive}
\title{Construct a new archive}
\usage{
archive(path)
}
\arguments{
\item{path}{File path to the archive.}
}
\description{
This function retrieves metadata about files in an archive, it can be passed
to `archive_con()` to create a connection to read a specific file from the
archive.
}
\examples{
a <- archive(system.file(package = "archive", "extdata", "data.zip"))
a
}
\seealso{
[archive_read()], [archive_write()] to read and write archive files
using R connections, [archive_extract()], [archive_write_files()],
[archive_write_dir()] to add or extract files from an archive.
}
