% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{archive_extract}
\alias{archive_extract}
\title{Extract contents of an archive}
\usage{
archive_extract(archive, dir = ".")
}
\arguments{
\item{archive}{An archive object or file path to the archive location.}

\item{dir}{Directory location to extract archive contents, will be created
if it does not exist.}
}
\description{
Extract contents of an archive
}
