% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_files.R
\name{archive_extract}
\alias{archive_extract}
\title{Extract contents of an archive}
\usage{
archive_extract(archive, dir = ".", file = NULL)
}
\arguments{
\item{archive}{\code{character(1)} The archive filename or an \code{archive} object.}

\item{dir}{\code{character(1)} Directory location to extract archive contents, will be created
if it does not exist.}

\item{file}{\code{character(1) || integer(1)} The filename within the archive,
specified either by filename or by position.}
}
\description{
Extract contents of an archive
}
\details{
If \code{file} is \code{NULL} (the default) all files will be extracted.
}
\examples{
a <- archive(system.file(package = "archive", "extdata", "data.zip"))
d <- tempfile()

# When called with default arguments extracts all files in the archive.
archive_extract(a, d)
list.files(d)
unlink(d)

# Can also specify one or more files to extract
d <- tempfile()
archive_extract(a, d, c("iris.csv", "airquality.csv"))
list.files(d)
unlink(d)
}
