% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{archive_read}
\alias{archive_read}
\title{Construct a read only connection into an archive file}
\usage{
archive_read(archive, file = 1L, mode = "r")
}
\arguments{
\item{archive}{An archive object or character vector to the archive}

\item{file}{The file to open the connection to. Can also be an numeric index
into the `archive()` data.frame.}

\item{mode}{The mode to open the file in.}
}
\description{
Construct a read only connection into an archive file
}
