% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_read.R
\name{archive_read}
\alias{archive_read}
\title{Create a readable connection to a file in an archive.}
\usage{
archive_read(archive, file = 1L, mode = "r", format = NULL, filter = NULL)
}
\arguments{
\item{mode}{\code{character(1)} A description of how to open the
connection (if it should be opened initially).  See section
‘Modes’ in \code{\link[base:connections]{base::connections()}} for possible values.}
}
\description{
Create a readable connection to a file in an archive.
}
\examples{
a <- archive(system.file(package = "archive", "extdata", "data.zip"))
# Show files in archive
a

# By default reads the first file in the archive.
read.csv(archive_read(a), nrows = 3)

# Can also specify a filename directly
read.csv(archive_read(a, "mtcars.csv"), nrows = 3)

# Or by position
read.csv(archive_read(a, 3), nrows = 3)

# Explicitly specify the format and filter if automatic detection fails.
read.csv(archive_read(a, format = "zip"), nrows = 3)
}
