% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{archive_write}
\alias{archive_write}
\alias{archive_write_dir}
\title{Construct a write only connection to a new archive}
\usage{
archive_write(archive, file, format = NULL, filter = NULL)

archive_write_dir(archive, dir, ..., recursive = TRUE, full.names = FALSE)
}
\arguments{
\item{archive}{The archive filename, if filter and format are both `NULL` the file extension will
be used to determine the archive format and filters, if any.}

\item{file}{The filename within the archive.}

\item{format}{`character(1)` default: `NULL` The archive format, one of \Sexpr[stage=render, results=rd]{archive:::choices_rd(names(archive:::archive_formats()))}.}

\item{filter}{`character(1)` default: `NULL` The archive filter, one of \Sexpr[stage=render, results=rd]{archive:::choices_rd(names(archive:::archive_filters()))}.}

\item{dir}{The directory of files to add}

\item{...}{additional paramters passed to `base::dir`}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}
}
\description{
Construct a write only connection to a new archive
}
