% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{archive_write_files}
\alias{archive_write_files}
\title{Add files to a new archive}
\usage{
archive_write_files(archive, files, format = NULL, filter = NULL)
}
\arguments{
\item{archive}{`character(1)` The archive filename.}

\item{files}{One or more files to add to the archive.}

\item{format}{`character(1)` default: `NULL` The archive format, one of \Sexpr[stage=render, results=rd]{archive:::choices_rd(names(archive:::archive_formats()))}.}

\item{filter}{`character(1)` default: `NULL` The archive filter, one of \Sexpr[stage=render, results=rd]{archive:::choices_rd(names(archive:::archive_filters()))}.}
}
\description{
`archive_write_files()` adds one or more files to a new archive.
`archive_write_dir()` adds all the file(s) in a directory to a new archive.
}
