% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{archive_write_files}
\alias{archive_write_files}
\title{Add files to a new archive}
\usage{
archive_write_files(archive, files)
}
\description{
`archive_write_files()` adds one or more files to a new archive.
`archive_write_dir()` adds all the file(s) in a directory to a new archive.
}
