% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{file_connections}
\alias{file_connections}
\alias{file_write}
\alias{file_read}
\title{Construct a file connections for (possibly compressed) files.}
\usage{
file_write(file, type)

file_read(file)
}
\arguments{
\item{file}{The filename within the archive.}

\item{type}{The type of filters to use for the file.}
}
\description{
This works similar to R's builtin [connections] for files. However it supports one on
or more of. `file_write()` returns an writable output connection,
`file_read()` returns a readable input connection.
}
\details{
- uuencode
- gzip compression
- bzip2 compression
- compress/LZW compression
- lzma, lzip, and xz compression
- lz4 compression
- lzop compression
}
