% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{file_connections}
\alias{file_connections}
\alias{file_write}
\alias{file_read}
\title{Construct a connections for (possibly compressed) files.}
\usage{
file_write(file, filter = NULL)

file_read(file, mode = "r")
}
\arguments{
\item{file}{\code{character(1) || integer(1)} The filename within the archive, specified either by filename or by position.}

\item{filter}{\code{character(1)} default: \code{NULL} The archive filter, one of \Sexpr[stage=render, results=rd]{archive:::choices_rd(names(archive:::archive_filters()))}.}
}
\description{
These work similar to R's built-in \link{connections} for files and differ from
\link{archive_read} and \link{archive_write} because they do not use an archive
format, just use one or more of the filters.
}
\details{
\code{file_write()} returns an writable output connection,
\code{file_read()} returns a readable input connection.
}
