% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{file_connections}
\alias{file_connections}
\alias{file_write}
\alias{file_read}
\title{Construct a file connections for (possibly compressed) files.}
\usage{
file_write(file, filter = NULL)

file_read(file)
}
\arguments{
\item{file}{The filename within the archive.}

\item{filter}{`character(1)` default: `NULL` The archive filter, one of \Sexpr[stage=render, results=rd]{archive:::choices_rd(names(archive:::archive_filters()))}.}
}
\description{
This works similar to R's builtin [connections] for files. However it
supports one on or more of the following
}
\details{
`file_write()` returns an writable output connection,
`file_read()` returns a readable input connection.
}
